code done of tasks

+++ app/Http/Controllers/Api/TaskController.php
@@ -4,6 +4,9 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Models\Task;
+use App\Models\Version;
+use Carbon\Carbon;
+use COM;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
@@ -25,30 +28,39 @@ class TaskController extends Controller
         }
         $userId =  Auth::user()->id;
         $task = $request->task;
-         
-        $Task = new Task();
-        $Task->userId = $userId;
-        $Task->task = $task;
-        $Task->save();
 
-        return response()->json([
-            'status'=> true,
-            'message' => 'Task Added successfully',
-            'data' => $Task
-        ], 200);
 
+        $findTask = Task::where('userId', $userId)->whereDate('created_at', date('Y-m-d'))->orderBy('id', 'desc')->first();
+        if (!$findTask) {
+
+            $Task = new Task();
+            $Task->userId = $userId;
+            $Task->task = $task;
+            $Task->save();
+
+            return response()->json([
+                'status' => true,
+                'message' => 'Task Added successfully',
+                'data' => $Task
+            ], 200);
+        } else {
+            return response()->json([
+                'status' => false,
+                'message' => 'Task Already Added',
+            ]);
+        }
     }
-    public function myTask ()
+    public function myTask()
     {
         $userId =  Auth::user()->id;
         $task = Task::where('userId', $userId)->orderBy('id', 'desc')->get();
         return response()->json([
-            'status'=> true,
+            'status' => true,
             'message' => 'Get Task successfully',
             'data' => $task
-        ],200);
+        ], 200);
     }
-    public function deleteTask (Request $request)
+    public function deleteTask(Request $request)
     {
         $validator = Validator::make($request->all(), [
             'id' => 'required'
@@ -61,14 +73,24 @@ class TaskController extends Controller
             ], 200);
         }
         $task = Task::find($request->id);
-        $task->delete();
-        return response()->json([
-            'status'=> true,
-            'message' => 'Task Deleted successfully',
-        ], 200);
+        
+        if ($task->created_at->format('Y-m-d') == date('Y-m-d')) {
 
+
+            $task->delete();
+
+            return response()->json([
+                'status' => true,
+                'message' => 'Task Deleted successfully',
+            ], 200);
+        } else {
+            return response()->json([
+                'status' => false,
+                'message' => 'you can not delete previous task',
+            ]);
+        }
     }
-    public function showTask (Request $request)
+    public function showTask(Request $request)
     {
         $validator = Validator::make($request->all(), [
             'id' => 'required',
@@ -83,12 +105,12 @@ class TaskController extends Controller
         }
         $task = Task::find($request->id);
         return response()->json([
-            'status'=> true,
+            'status' => true,
             'message' => 'Task List successfully',
             'data' => $task
         ], 200);
     }
-    public function updateTask (Request $request)
+    public function updateTask(Request $request)
     {
         $validator = Validator::make($request->all(), [
             'id' => 'required',
@@ -101,40 +123,58 @@ class TaskController extends Controller
                 'errors' => $validator->errors()
             ], 200);
         }
+
         $task = Task::find($request->id);
-        $task->task = $request->task;
-        $task->save();
-        return response()->json([
-            'status'=> true,
-            'message' => 'Task Updated successfully',
-            'data' => $task
-        ], 200);
+       
+        if ($task->created_at->format('Y-m-d') == date('Y-m-d')) {
+
+            $task->task = $request->task;
+            $task->save();
+            return response()->json([
+                'status' => true,
+                'message' => 'Task Updated successfully',
+                'data' => $task
+            ], 200);
+        } else {
+            return response()->json([
+                'status' => false,
+                'message' => 'you can only update today task',
+            ]);
+        }
     }
     public function allTask(Request $request)
     {
         $date = $request->date;
         $name = $request->name;
-    
+
         $query = Task::orderBy('id', 'desc')->with('taskWithUser');
-    
-        if($name){
-            $query->whereHas('taskWithUser', function($q) use ($name) {
+
+        if ($name) {
+            $query->whereHas('taskWithUser', function ($q) use ($name) {
                 $q->where('name', 'like', '%' . $name . '%');
             });
         }
-    
-        if($date){
+
+        if ($date) {
             $query->whereDate('created_at', $date);
         }
-    
+
         $task = $query->get();
-    
+
         return response()->json([
             'status' => true,
             'message' => 'Task List successfully',
             'data' => $task
         ], 200);
     }
-    
-}
+    public function version()
+    {
+        $version = Version::orderBy('id', 'desc')->first();
 
+        return response()->json([
+            'status' => true,
+            'message' => 'Version List successfully',
+            'data' => $version
+        ], 200);
+    }
+}
